backend_django/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
mysql-connector-python>=8.0
python-decouple>=3.8
pymysql>=1.1.0